refactor(ecau): consolidate types in single file
