+++ src/mb_enhanced_cover_art_uploads/providers/allmusic.ts
@@ -1,6 +1,6 @@
 import { safeParseJSON } from '@lib/util/json';
 
-import type { CoverArt } from './base';
+import type { CoverArt } from '../types';
 import { CoverArtProvider } from './base';
 
 interface AllMusicImage {
