+++ src/mb_enhanced_cover_art_uploads/providers/bandcamp.ts
@@ -6,7 +6,8 @@ import { filterNonNull } from '@lib/util/array';
 import { parseDOM, qs, qsa, qsMaybe } from '@lib/util/dom';
 import { getImageDimensions } from '@src/mb_caa_dimensions/dimensions';
 
-import type { CoverArt, ParsedTrackImage } from './base';
+import type { CoverArt } from '../types';
+import type { ParsedTrackImage } from './base';
 import { ProviderWithTrackImages } from './base';
 
 export class BandcampProvider extends ProviderWithTrackImages {
