+++ src/mb_enhanced_cover_art_uploads/providers/vgmdb.ts
@@ -6,7 +6,7 @@ import { safeParseJSON } from '@lib/util/json';
 import { urlBasename } from '@lib/util/urls';
 import { gmxhr } from '@lib/util/xhr';
 
-import type { CoverArt } from './base';
+import type { CoverArt } from '../types';
 import { CoverArtProvider } from './base';
 
 // Not full, only what we need
