+++ src/mb_enhanced_cover_art_uploads/types.ts
@@ -1,0 +1,64 @@
+// Common type definitions
+
+import type { ArtworkTypeIDs } from '@lib/MB/CoverArt';
+
+export interface CoverArtOptions {
+    /**
+     * Artwork types to set. May be empty or undefined.
+     */
+    types?: ArtworkTypeIDs[];
+    /**
+     * Comment to set. May be empty or undefined.
+     */
+    comment?: string;
+}
+
+/** Cover art provided by a provider. */
+export interface CoverArt extends CoverArtOptions {
+    /**
+     * URL to fetch.
+     */
+    url: URL;
+    /**
+     * Whether maximisation should be skipped for this image. If undefined,
+     * interpreted as false.
+     */
+    skipMaximisation?: boolean;
+}
+
+/** Cover art that still needs to be checked before it can be fetched, i.e. URLs provided by the user. */
+export interface BareCoverArt extends CoverArtOptions {
+    url: URL;
+}
+
+/** Contents of a fetched image. */
+export interface ImageContents {
+    requestedUrl: URL;
+    fetchedUrl: URL;
+    wasRedirected: boolean;
+    file: File;
+}
+
+interface BaseFetchedImage {
+    originalUrl: URL;
+    maximisedUrl: URL;
+    fetchedUrl: URL;
+    wasMaximised: boolean;
+    wasRedirected: boolean;
+}
+
+/** Image that was fetched. */
+export interface FetchedImage extends BaseFetchedImage {
+    content: File;
+    // types and comment may be empty or undefined. If undefined, the value
+    // will be replaced by the default, if any. If defined but empty, the
+    // default will not be used.
+    types?: ArtworkTypeIDs[];
+    comment?: string;  // Can be empty string
+}
+
+/** Batch of images that were fetched, possibly from a provider. */
+export interface FetchedImageBatch {
+    images: FetchedImage[];
+    containerUrl?: URL;
+}
